rename: Loader > Skeleton으로 변경
